Refactor
All hooks related with offers moved to one file named 'offers'
